+++ app/subscribed/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import PodcastCard from "@/components/PodcastCard";
 import PodcastDisplay from "@/components/PodcastDisplay";
-import { zMetaData } from "@/types/podcastTypes";
+import { zMetaData } from "@/@types/podcastTypes";
 import { useQuery } from "@tanstack/react-query";
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/navigation";
