+++ components/PodcastDisplay.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { MetaData } from "@/types/podcastTypes";
+import { MetaData } from "@/@types/podcastTypes";
 import React from "react";
 import PodcastCard, { variant } from "./PodcastCard";
 import { cn } from "@/lib/utils";
