+++ components/player/EpisodeImage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { AspectRatio } from "../ui/aspect-ratio";
 import Image from "next/image";
-import { EpisodeData } from "@/types/podcastTypes";
+import { EpisodeData } from "@/@types/podcastTypes";
 import { z } from "zod";
 import { cn } from "@/lib/utils";
